chore: refactor text component
