+++ src/app/shared/components/template/components/text/text.component.html
@@ -1,0 +1,14 @@
+@if (this._row.value) {
+  <div
+    class="large standard normal margin-t-large"
+    [class]="params.style"
+    [ngStyle]="isFalsy ? { display: 'none' } : { display: 'block' }"
+    [innerHTML]="
+      params.type === 'numbered'
+        ? (this._row.value | number)
+        : (this._row.value?.toString() | markdown)
+    "
+    [style]="_row.style_list | styleList"
+    [style.textAlign]="params.textAlign"
+  ></div>
+}
